165716787-feature(home page): refactor code
